bw: implement NEST_SET_DOOR

+++ BinWeevils.Protocol/Str/NestJoinLocRequest.cs
@@ -5,7 +5,7 @@ namespace BinWeevils.Protocol.Str
     public partial record struct NestJoinLocRequest
     {
         [StrField] public int m_locID;
-        [StrField] public uint m_doorID;
+        [StrField] public byte m_doorID;
         [StrField] public double m_x;
         [StrField] public double m_y;
         [StrField] public double m_z;
