+++ BinWeevils.Protocol/Str/NestJoinedLocNotification.cs
@@ -4,7 +4,7 @@ namespace BinWeevils.Protocol.Str
 {
     public partial record struct NestJoinedLocNotification
     {
-        [StrField] public int m_userID;
+        [StrField] public ulong m_userID;
         [StrField] public NestJoinLocRequest m_body;
     }
 }
